Start adding a new `createAnalysis` method
This is a substitute for the analysisManager's `runAnalysis` method. The
big difference is that this thing returns an object with some async
calls which emulate the exact backend API that we have for Analyses.

So, for instance, you can call

```
const analysis = await createAnalysis(params);
analysis.addLocation(location);
analysis.findPoints();
analysis.runAnalysis();
analysis.releaseAnalysis();
```

And hey! You're programmatic API looks just like the actual Protocol API
that it is exercising, meaning you have full control over which parts of
the process get run and which parts get skipped, and you can even do
things in-between the calls, like checking how many points you received,
and choosing whether or not to actually run that analysis.